fixing data type issue

+++ BaseResNetRunner.py
@@ -41,7 +41,7 @@ class BaseResNetRunner(BaseRunner):
     def test_batch_and_get_metrics(self, batch):
         #forward pass
         pred = self.do_forward_pass(batch)
-        loss = self.loss_fn(pred, batch[2])
+        loss = self.loss_fn(pred, batch[2].float())
         acc  = self.get_accuracy(pred, batch[2])
 
         return [('loss', loss.mean().item()), ('acc', acc)]
